requirements.txt
asgiref==3.9.1
Django==5.2.4
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
sqlparse==0.5.3
